refactor : Validator Running 클래스에서 InputView 클래스로 이동
