+++ src/main/java/basball/baseball/Running.java
@@ -40,33 +40,5 @@ public class Running {
         }
     }
 
-    public static class Validator {
-        private static final Pattern PATTERN = Pattern.compile("[1-9]{3}");
-
-        public static void hasThreeNumber(String next) {
-
-            if (!PATTERN.matcher(next).matches()) {
-                throw new IllegalArgumentException("서로 다른 3자리 숫자만 입력 해주세요");
-            }
-        }
-
-        public static void isDifferentNumbers(int[] numbers) {
-
-            if (isSameNumber(numbers[0], numbers[1]) || isSameNumber(numbers[0], numbers[2]) || isSameNumber(numbers[1], numbers[2])) {
-                throw new IllegalArgumentException("서로 다른 수를 입력해주세요");
-            }
-        }
-
-        private static boolean isSameNumber(int number1, int number2) {
-            return number1 == number2;
-        }
-
-        public static void validateAnswer(String answer) {
-            if (!answer.equals("1") && !answer.equals("2")) {
-                throw new IllegalArgumentException("1 또는 2만 입력 해 주세요");
-            }
-        }
-    }
-
 }
