+++ src/main/java/basball/baseball/model/Balls.java
@@ -1,16 +1,11 @@
 package basball.baseball.model;
 
-import basball.baseball.Running;
-import org.jetbrains.annotations.NotNull;
-
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
-import java.util.stream.Collectors;
 
-import static basball.baseball.Running.Validator.*;
 import static basball.baseball.model.Balls.NumberGenerator.generate;
+import static basball.view.InputView.Validator.isDifferentNumbers;
 
 public class Balls {
     private static final int COUNT = 3;
